test/caesar.test.ts
import {encrypt, decrypt} from '../src/caesar';

describe('Caesar cipher - encryption', () => {
  test('Shift', () => {
    expect(encrypt('abcdefghijklmnopqrstuvwxyz', 1)).toEqual(
      'bcdefghijklmnopqrstuvwxyza'
    );
    expect(encrypt('abcdefghijklmnopqrstuvwxyz', -1)).toEqual(
      'zabcdefghijklmnopqrstuvwxy'
    );
  });
  test('Case sensitive', () => {});
  test('Include foreign chars', () => {});
});

describe('Caesar cipher - decryption', () => {
  test('Shift', () => {
    expect(decrypt('bcdefghijklmnopqrstuvwxyza', 1)).toEqual(
      'abcdefghijklmnopqrstuvwxyz'
    );
    expect(decrypt('bcd', -1)).toEqual('cde');
  });
  test('Case sensitive', () => {});
  test('Include foreign chars', () => {});
});
